refactor(f): sanity lib organization
