+++ packages/frontend/src/lib/sanity/builder.ts
@@ -1,48 +1,7 @@
 /* eslint-disable @typescript-eslint/restrict-template-expressions */
-import { type ClientConfig, createClient } from '@sanity/client'
-import { createImageUrlBuilder, type SanityImageSource } from '@sanity/image-url'
-import type { PortableTextBlock, PortableTextSpan } from '@portabletext/types'
 
-// Environment variables, found in ".env". Check ".env.example" for explanation.
-import type { Query } from './schema'
-import { dev } from '$app/environment'
-import { isPortableTextSpan } from '@sanity/types'
-
-// if (!SANITY_PROJECT_ID || !SANITY_DATASET) {
-// 	throw new Error('Did you forget to run yarn run -T sanity init --env?');
-// }
-
-// if (!import.meta.env.PUBLIC_SANITY_API_VERSION) {
-// 	throw new Error('Did you forget to add an API Version environment variable?');
-// }
-
-// It's okay to expose projectId
-// See: https://www.sanity.io/answers/hello-quick-question-is-it-safe-to-commit-p1609342625280000
-const config: ClientConfig = {
-  projectId: '1ah7xxlt',
-  dataset: 'production',
-  useCdn: true,
-  apiVersion: '2024-09-20',
-}
-
-// Change the dataset if it is a development environment.
-if (dev || import.meta.env.MODE === 'development') {
-  config.dataset = 'development'
-  config.useCdn = false
-} else if (import.meta.env.MODE === 'staging') {
-  config.dataset = 'staging'
-  config.useCdn = false
-}
-
-export const client = createClient(config)
-
-// Helps transform images from Sanity.
-// See: https://www.sanity.io/docs/presenting-images#mY9Be3Ph
-const builder = createImageUrlBuilder(client)
-
-export function urlFor(source: SanityImageSource) {
-  return builder.image(source)
-}
+import type { Query } from '../schema'
+import { client } from './index'
 
 /**
  * Makes a request to the Sanity API with the given GROQ query string.
@@ -195,31 +154,4 @@ export function getAttrs(attrs: string[]): string {
  */
 export function getOrder(order: string): string {
   return `| order(${order})`
-}
-
-/**
- * `blocksToText` strings together portable text spans into one string. If content doesn't exist,
- * it just returns an empty string. Slightly modified from:
- * https://www.sanity.io/docs/developer-guides/presenting-block-text
- * @param content
- */
-export function blocksToText(content?: PortableTextBlock[]) {
-  return (
-    content
-      ?.map(block => {
-        // This `if` block fixes issue #337. Also, no second check because its always falsy.
-        if (block._type !== 'block' /* && !block.children */) {
-          return ''
-        }
-        return (
-          block.children
-            // Technically some generic typing is not needed to be so explicit, but it makes things
-            // more obvious to read in this short portable text pipeline.
-            .filter<PortableTextSpan>(child => isPortableTextSpan(child))
-            .map<string>(child => child.text)
-            .join('')
-        )
-      })
-      .join('\n\n') ?? ''
-  )
 }
