+++ packages/frontend/src/routes/(app)/series/+page.server.ts
@@ -1,5 +1,5 @@
 import { error } from '@sveltejs/kit'
-import { buildSanityQuery, sanityFetch } from '$lib/sanity'
+import { buildSanityQuery, sanityFetch } from '$lib/sanity/builder'
 import type { PageServerLoad } from './$types'
 import type { Series } from '$lib/schema'
 import type { MetaTagsProps } from 'svelte-meta-tags'
